EMP-8: change self_delay to work with 1 or 2 bytes

+++ tokens/tokens.h
@@ -237,7 +237,7 @@ void build_masked_tokens_cust(
   struct Balance_l val_cpfp,
   struct Balance_l bal_min_cust,
   struct Balance_l bal_min_merch,
-  uint32_t self_delay,
+  uint16_t self_delay,
 
   struct CommitmentRandomness_l revlock_commitment_randomness_l,
   struct State_l w_new,
@@ -308,7 +308,7 @@ void build_masked_tokens_merch(
   struct Balance_l val_cpfp,
   struct Balance_l bal_min_cust,
   struct Balance_l bal_min_merch,
-  uint32_t self_delay,
+  uint16_t self_delay,
 
   struct HMACKey_l hmac_key,
   struct Mask_l merch_mask_l,
